refactor: input component structure
